more flexibility for linux

+++ setup.sh
@@ -24,9 +24,15 @@ vim +PlugInstall +qall
 
 if ! command -v fzf &> /dev/null
 then
-  # FZF
-  git clone --depth 1 https://github.com/junegunn/fzf.git "$HOME/.fzf"
-  $HOME/.fzf/install
+  if [[ "$platform" == "linux" ]]
+  then
+    sudo apt update
+    sudo apt install fzf
+  else 
+    # FZF
+    git clone --depth 1 https://github.com/junegunn/fzf.git "$HOME/.fzf"
+    $HOME/.fzf/install
+  fi
 fi
 
 if [ ! -d "$HOME/bin" ]
@@ -37,15 +43,21 @@ fi
 # Ripgrep
 if ! command -v rg &> /dev/null
 then
-  curl -LO 'https://github.com/BurntSushi/ripgrep/releases/download/11.0.2/ripgrep-11.0.2-x86_64-unknown-linux-musl.tar.gz'
-  tar xf ripgrep-11.0.2-*.tar.gz
-  if [ ! -d "$HOME/bin/rg" ]
+  if [[ "$platform" == "linux" ]]
   then
-      mkdir "$HOME/bin/rg"
+    sudo apt update
+    sudo apt install ripgrep
+  else 
+    curl -LO 'https://github.com/BurntSushi/ripgrep/releases/download/11.0.2/ripgrep-11.0.2-x86_64-unknown-linux-musl.tar.gz'
+    tar xf ripgrep-11.0.2-*.tar.gz
+    if [ ! -d "$HOME/bin/rg" ]
+    then
+        mkdir "$HOME/bin/rg"
+    fi
+    cp ripgrep-11.0.2-*/rg "$HOME/bin/rg"
+    rm ripgrep-11.0.2-*.tar.gz && rm -rf ripgrep-11.0.2-x86_64*
+    echo "export PATH=\$PATH:\$HOME/bin/rg" >> ~/.bashrc
   fi
-  cp ripgrep-11.0.2-*/rg "$HOME/bin/rg"
-  rm ripgrep-11.0.2-*.tar.gz && rm -rf ripgrep-11.0.2-x86_64*
-  echo "export PATH=\$PATH:\$HOME/bin/rg" >> ~/.bashrc
 fi
 
 # silver searcher
